+++ src/main/java/racingCar/domain/moveStrategy/MoveBehavior.java
@@ -1,4 +1,4 @@
-package racingCar.domain;
+package racingCar.domain.moveStrategy;
 
 public interface MoveBehavior {
     boolean isMove();
