+++ src/test/java/racingCar/domain/RacingGameTest.java
@@ -1,10 +1,9 @@
-package racingCar.application;
+package racingCar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingCar.domain.RacingCar;
 
 import java.util.Arrays;
 import java.util.List;
